add guards to gcloud functions for when gcloud has been disabled (no key)

+++ server/services/gcloud/tasks.go
@@ -43,6 +43,11 @@ func InitTasks() func() {
 }
 
 func CreateEmailTask(email string, ownerName string, eventName string, eventId string) []string {
+	if TasksClient == nil {
+		logger.StdErr.Println("WARNING: Cloud Tasks is disabled, skipping CreateEmailTask")
+		return []string{}
+	}
+
 	// Get listmonk url env vars
 	listmonkUrl := os.Getenv("LISTMONK_URL")
 	listmonkUsername := os.Getenv("LISTMONK_USERNAME")
@@ -131,6 +136,11 @@ func CreateEmailTask(email string, ownerName string, eventName string, eventId s
 }
 
 func DeleteEmailTask(taskId string) {
+	if TasksClient == nil {
+		logger.StdErr.Println("WARNING: Cloud Tasks is disabled, skipping DeleteEmailTask")
+		return
+	}
+
 	err := TasksClient.DeleteTask(context.Background(), &cloudtaskspb.DeleteTaskRequest{
 		Name: taskId,
 	})
